update android config for the 1.24.0
add new i2c applets, truncate and uevent

tested on aosp M x86_64 cm-11.0 (ics) and cm-13.0 arm (bacon)

i2c headers are present for pre-21 android build platforms (like ics)

+++ archival/gzip.c
@@ -96,7 +96,7 @@ aa:      85.1% -- replaced with aa.gz
 
 #include "libbb.h"
 #include "bb_archive.h"
-
+#include <strings.h> // ffs()
 
 /* ===========================================================================
  */
@@ -748,7 +748,7 @@ static int longest_match(IPos cur_match)
 		if (len > best_len) {
 			G1.match_start = cur_match;
 			best_len = len;
-			if (len >= nice_match)
+			if (len >= (int) nice_match)
 				break;
 			scan_end1 = scan[best_len - 1];
 			scan_end = scan[best_len];
